+++ src/main/java/browser/FirefoxTest.java
@@ -1,4 +1,4 @@
-package browser.subfolder;
+package browser;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
