+++ src/main/java/browser/SafariTest.java
@@ -1,4 +1,4 @@
-package browser.subfolder;
+package browser;
 
 public class SafariTest {
 //    SafariDriver is only supported on macOS.
