--- src/main/java/browser/Test.java
@@ -1,4 +1,0 @@
-package browser;
-
-public class Test {
-}
